fixed review comments; refactored state changing
