+++ Roman_Balashov/react_project/package.json
@@ -5,8 +5,8 @@
   "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack --mode=production",
-    "server": "webpack-dev-server --mode=development"
+    "build": "webpack --mode=production",
+    "dev": "webpack-dev-server --mode=development"
   },
   "author": "Roman Balashov",
   "license": "ISC",
