+++ Roman_Balashov/react_project/src/components/TextField.jsx
@@ -1,11 +1,7 @@
 import React from 'react';
 
-// const TextField = () => <input type="text" className="input" />;
-
-class TextField extends React.Component {
+export class TextField extends React.Component {
     render() {
         return <input type="text" className="input" />;
     }
 }
-
-export default TextField;
