+++ Roman_Balashov/react_project/webpack.config.js
@@ -17,17 +17,14 @@ module.exports = {
                     loader: 'babel-loader',
                     options: {
                         presets: ['@babel/env', '@babel/react'],
-                        plugins: [
-                            [
-                                "@babel/plugin-proposal-class-properties",
-                                {
-                                    "loose": true
-                                }
-                            ]
-                    }
+                        plugins: ['@babel/plugin-proposal-class-properties'],
+                    },
                 },
             },
         ],
+    },
+    resolve: {
+        extensions: ['.jsx', '.js']
     },
     plugins: [
         new HTMLWebpackPlugin({ template: path.resolve(__dirname, "src", "index.html"), }),
